Relying on input ctx to cancel ssm session, splitting out ecs session
